Feature/31 2,31 3,37 2/all/login be

+++ backend/testhelper/src/main/java/kr/ac/ajou/da/testhelper/account/AccountMapper.java
@@ -1,15 +1,15 @@
 package kr.ac.ajou.da.testhelper.account;
 
-import java.sql.SQLException;
-
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 import org.apache.ibatis.annotations.Select;
 import org.springframework.stereotype.Repository;
 
+import java.sql.SQLException;
+
 @Mapper
 @Repository
 public interface AccountMapper {
 	@Select("SELECT role FROM ACCOUNT WHERE id = ${accountId}")
-	String getAccountRole(@Param("accountId") int accountId) throws SQLException;
+	String getAccountRole(@Param("accountId") Long accountId) throws SQLException;
 }
